+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to
 ### Added
 
 - ✨(backend) add a is_first_connection flag to the User model#1938
+### Changed
+
+- ♿(frontend) focus skip link on headings and skip grid dropzone #1983
 
 ## [v4.7.0] - 2026-03-09